Moved code for coreir libraries to its own file
